--- frontend/.storybook/preview.ts
@@ -1,15 +1,0 @@
-import type { Preview } from '@storybook/react-webpack5';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  tags: ['autodocs'],
-};
-
-export default preview;
