+++ frontend/src/components/Header/Header.stories.tsx
@@ -1,8 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import Header from './Header';
-import { Global, ThemeProvider } from '@emotion/react';
-import { theme } from '../../styles/theme';
-import reset from '../../styles/reset';
 
 const meta: Meta<typeof Header> = {
   title: 'common/Header',
@@ -16,19 +13,9 @@ export default meta;
 type Story = StoryObj<typeof Header>;
 
 export const Default: Story = {
-  render: () => (
-    <ThemeProvider theme={theme}>
-      <Global styles={reset} />
-      <Header activeNav="home" />
-    </ThemeProvider>
-  ),
+  render: () => <Header activeNav="home" />,
 };
 
 export const Recommend: Story = {
-  render: () => (
-    <ThemeProvider theme={theme}>
-      <Global styles={reset} />
-      <Header activeNav="recommend" />
-    </ThemeProvider>
-  ),
+  render: () => <Header activeNav="recommend" />,
 };
